+++ plugins/postcss-gradients-interpolation-method/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Gradients Interpolation Method
 
+### Unreleased (patch)
+
+- fix dependencies
+
 ### 1.0.0 (March 4, 2022)
 
 - Initial version